+++ infra/reachability-metadata/src/main/resources/META-INF/native-image/com.atomikos/transactions/6.0.0/resource-config.json
@@ -1,0 +1,8 @@
+{
+  "resources":{
+  "includes":[{
+    "condition":{"typeReachable":"com.atomikos.icatch.provider.imp.AssemblerImp"},
+    "pattern":"\\Qtransactions-defaults.properties\\E"
+  }]},
+  "bundles":[]
+}
