+++ infra/reachability-metadata/src/main/resources/META-INF/native-image/com.h2database/h2/2.2.224/reflect-config.json
@@ -1,0 +1,23 @@
+[
+{
+  "condition":{"typeReachable":"org.h2.jdbcx.JdbcDataSource"},
+  "name":"org.h2.jdbcx.JdbcDataSource",
+  "allDeclaredClasses": true,
+  "allDeclaredMethods": true,
+  "allDeclaredFields": true,
+  "allDeclaredConstructors": true,
+  "allPublicClasses": true,
+  "allPublicMethods": true,
+  "allPublicFields": true,
+  "allPublicConstructors": true,
+  "allRecordComponents": true,
+  "allNestMembers": true,
+  "allSigners": true,
+  "allPermittedSubclasses": true,
+  "queryAllDeclaredMethods": true,
+  "queryAllDeclaredConstructors": true,
+  "queryAllPublicMethods": true,
+  "queryAllPublicConstructors": true,
+  "unsafeAllocated": true
+}
+]
