+++ pom.xml
@@ -1063,7 +1063,6 @@
                             <extensions>true</extensions>
                             <configuration>
                                 <buildArgs>
-                                    <buildArg>-J-Xmx7g</buildArg>
                                     <buildArg>-H:+AddAllCharsets</buildArg>
                                 </buildArgs>
                                 <quickBuild>true</quickBuild>
