+++ test/native/native-image-filter/extra-filter.json
@@ -10,12 +10,15 @@
     {"includeClasses": "java.util.Properties"},
     {"excludeClasses": "javax.security.auth.x500.**"},
     {"excludeClasses": "javax.smartcardio.**"},
+    {"excludeClasses": "javax.xml.stream.**"},
     {"excludeClasses": "jdk.internal.misc.**"},
     {"excludeClasses": "sun.misc.**"},
     {"excludeClasses": "sun.nio.ch.**"},
     {"excludeClasses": "sun.security.**"},
 
     {"excludeClasses": "android.app.**"},
+    {"excludeClasses": "com.arjuna.**"},
+    {"excludeClasses": "com.atomikos.logging.**"},
     {"excludeClasses": "com.fasterxml.jackson.databind.**"},
     {"excludeClasses": "com.github.benmanes.caffeine.cache.**"},
     {"excludeClasses": "com.github.dockerjava.api.**"},
@@ -24,20 +27,24 @@
     {"excludeClasses": "com.microsoft.sqlserver.jdbc.**"},
     {"excludeClasses": "com.mysql.cj.**"},
     {"excludeClasses": "com.zaxxer.hikari.**"},
+    {"excludeClasses": "ch.qos.logback.classic.**"},
     {"excludeClasses": "io.grpc.**"},
     {"excludeClasses": "io.netty.**"},
     {"excludeClasses": "io.vertx.core.**"},
     {"excludeClasses": "groovy.**"},
     {"excludeClasses": "libcore.io.**"},
     {"excludeClasses": "org.apache.calcite.**"},
+    {"excludeClasses": "org.apache.logging.log4j.**"},
+    {"excludeClasses": "org.apache.log4j.**"},
     {"excludeClasses": "org.apache.zookeeper.**"},
     {"excludeClasses": "org.codehaus.groovy.**"},
     {"excludeClasses": "org.h2.**"},
+    {"excludeClasses": "org.jboss.logmanager.**"},
     {"excludeClasses": "org.locationtech.jts.geom.**"},
     {"excludeClasses": "org.opengauss.**"},
     {"excludeClasses": "org.postgresql.**"},
     {"excludeClasses": "org.robolectric.**"},
-    {"excludeClasses": "org.slf4j.event.**"},
+    {"excludeClasses": "org.slf4j.**"},
     {"excludeClasses": "org.testcontainers.**"}
   ],
   "regexRules": [
