+++ test/native/src/test/resources/test-native/yaml/modes/cluster/etcd.yaml
